+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 This is a manually generated log to track changes to the repository for each release. 
 Each section should include general headers such as **Implemented enhancements** 
 and **Merged pull requests**. All closed issued and bug fixes should be 
-represented by the pull requests that fixed them. This log originated with Singularity 2.4
-and changes prior to that are (unfortunately) done retrospectively. 
+represented by the pull requests that fixed them. 
 Critical items to know are:
 
  - renamed commands
@@ -18,5 +17,6 @@ The client here will eventually be released as "spython" (and eventually to
 singularity on pypi), and the versions here will coincide with these releases.
 
 ## [master](https://github.com/singularityhub/singularity-cli/tree/master)
+ - addition of docker and singularity recipe parsers (0.0.21)
  - generating basic api documentation (sphinx) (0.0.2)
  - adding changelog, and original code for client  (0.0.1)